+++ src/pages/Home/Hero.jsx
@@ -1,9 +1,8 @@
 import React from "react";
-import Hero_image from "../images/Hero_img.png";
+import Hero_image from "images/Hero_img.png";
 
 import Button from "@mui/material/Button";
 import Box from "@mui/material/Box";
-import Companies from "./Companies";
 
 const commonStyles = {
     bgcolor: "#D9D9D9",
